Fix issue with negative hours (in COs) not being handled
Change orders with negative labour hours were not being written to
worksheets, since the Labour property looked only for items that had
labour qty > 0.  However, when we are removing labour in a CO, these
items would be skipped, even though they are actually labour items.

The test was changed to look for non-zero qty rather than >0